+++ PlantUmlViewer/PlantUmlViewer/Forms/PreviewWindow.Designer.cs
@@ -34,7 +34,7 @@
             this.loadingCircleToolStripMenuItem_Refreshing = new MRG.Controls.UI.LoadingCircleToolStripMenuItem();
             this.toolStripStatusLabel_Time = new System.Windows.Forms.ToolStripStatusLabel();
             this.toolStripStatusLabel_Zoom = new System.Windows.Forms.ToolStripStatusLabel();
-            this.imageBox_Diagram = new Cyotek.Windows.Forms.ImageBox();
+            this.imageBox_Diagram = new Forms.Controls.CustomImageBox();
             this.contextMenuStrip_Diagram = new System.Windows.Forms.ContextMenuStrip(this.components);
             this.ToolStripMenuItem_Diagram_CopyToClipboard = new System.Windows.Forms.ToolStripMenuItem();
             this.ToolStripMenuItem_Diagram_ExportFile = new System.Windows.Forms.ToolStripMenuItem();
@@ -408,7 +408,7 @@
 
         private System.Windows.Forms.TableLayoutPanel tableLayoutPanel_Window;
         private System.Windows.Forms.Button button_Refresh;
-        private Cyotek.Windows.Forms.ImageBox imageBox_Diagram;
+        private Forms.Controls.CustomImageBox imageBox_Diagram;
         private System.Windows.Forms.Button button_Export;
         private System.Windows.Forms.StatusStrip statusStrip_Bottom;
         private System.Windows.Forms.ToolStripStatusLabel toolStripStatusLabel_Zoom;